verification/requirements.txt
flask==3.0.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
imagehash==4.3.1
scikit-image==0.22.0
exifread==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
